requirements.txt
# Core
torch
numpy
pandas
matplotlib
scikit-learn

# Vision / video
opencv-python; python_version < "3.13"
decord
pillow
imageio
imageio-ffmpeg

# Models
transformers
ultralytics          # YOLOv8

# Cube logic
pycuber
kociemba; python_version < "3.13"

# UI
streamlit

# Dev
pytest
